% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epivizrStandalone-package.r
\docType{package}
\name{epivizrStandalone}
\alias{epivizrStandalone}
\alias{epivizrStandalone-package}
\alias{getStandaloneLocation}
\title{epivizrStandalone.}
\usage{
getStandaloneLocation()
}
\description{
epivizrStandalone.
}

