% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epivizrStandalone-package.r
\name{initStandalone}
\alias{initStandalone}
\title{Initialize Epiviz Standalone to either a local epiviz repository or clones from github.}
\usage{
initStandalone(url = "https://github.com/epiviz/epiviz.git",
  branch = "master", local_path = NULL)
}
\arguments{
\item{url}{(character) github url to use. defaults to (\url{"https://github.com/epiviz/epiviz.git"}).}

\item{branch}{(character) branch on the github repository. defaults to (master).}

\item{local_path}{(character) if you already have a local instance of epiviz and would like to run standalone use this.}
}
\value{
path to the standalone location
}
\description{
Initialize Epiviz Standalone to either a local epiviz repository or clones from github.
}
\examples{
see package vignete for example usage
initStandalone(url="https://github.com/epiviz/epiviz.git", branch="master")

}

