% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epivizrStandalone-package.r
\name{startStandalone}
\alias{startStandalone}
\title{Uses the local instance of epiviz to start a stanadalone epivizr session}
\usage{
startStandalone()
}
\value{
An object of class \code{\link{EpivizApp}}
}
\description{
Uses the local instance of epiviz to start a stanadalone epivizr session
}
\examples{
see package vignete for example usage
app <- startStandalone()
app$stop_app()

}

