% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startStandalone.R
\name{startStandalone}
\alias{startStandalone}
\title{Start a standalone \code{epivizr} session.}
\usage{
startStandalone(...)
}
\arguments{
\item{...}{additional arguments passed to \code{\link[epivizr]{startEpiviz}}.}
}
\value{
An object of class \code{\link[epivizr]{EpivizApp}}
}
\description{
Uses the local repository of epiviz JS app to start a standalone epivizr session
through the \code{\link[epivizr]{startEpiviz}} function.
}
\examples{
# see package vignete for example usage
app <- startStandalone(non_interactive=TRUE, open_browser=TRUE)
app$stop_app()

}

